requirements.txt
Flask==3.0.3
feedparser==6.0.11
newspaper3k==0.2.8
nltk==3.9.1
Pillow==10.4.0
matplotlib==3.9.2
lxml_html_clean==0.3.1
wordcloud==1.9.3
cairosvg==2.7.1